Separated player API and admin
